# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления личной музыкальной коллекцией",
        "title": "Music Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/statistics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Статистика медиатеки",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tracks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Получить все треки",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Media"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Добавить трек",
                "parameters": [
                    {
                        "description": "Данные трека",
                        "name": "track",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Media"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Media"
                        }
                    }
                }
            }
        },
        "/tracks/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Поиск треков",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название или исполнитель",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Media"
                            }
                        }
                    }
                }
            }
        },
        "/tracks/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Обновить трек",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID трека",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "track",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Media"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Media"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tracks"
                ],
                "summary": "Удалить трек",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID трека",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Media": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "track": {
                    "type": "string"
                },
                "track_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}